Removed the FilterValidator interface, because it doesn't add any value. Renamed the FilterValidatorImpl class to FilterValidator.
